requirements.txt
pytest
requests
python-dotenv
allure-pytest
psycopg2-binary
